Apostrophe and Middle Name Update
Added the ability to use apostrophes in the titles of books and the names of publishers.

Added an optional middle name box for authors names. This will allow for better formatting with the names of the authors.

Had to do a lot of reworking on the add.fxml file to make the middle name fit.