Book Attribute class added
Created the 'BookAttributes' class to handle all the observable lists and bidi hashmaps that the program uses.

This required an extensive rework to the 'AddBookController' (which orginally was the only class that used these variables) so that it could still access them as needed.

Also improved support for middle names in author names.