Details Pane Edit Functionality
Added the needed logic, GUI elements and query to allow for the program to be able to edit the information for existing books in the database.

This required an extensive re-working & additions to the code of the 'detailsControlller'  class and 'details_edit' file. There are now several overlapping GUI elements that swap out with each other depending on what role the pane is performing at that particular moment. Most of the existing GUI elements were also replaced with auto-complete textfields to facilitate this.

One query for updating the book information was added to the 'SQLCommands' class. Another class was edited to select the genre_type element for a specific book.