Extensive bug fixes
A variety of issues occured after the last push.
The last push wasn't tested thoroughly enough and was actually quite broken.
The biggest issues were, in order:
1. Major slowdowns when opening and reopening the 'add book' pane.
2. Unable to add new authors, publishers, genres etc...
3. Errors when doing a variety of other misc actions throughout the code.

The alterations and testing that was done to fix this issue are too extensive to list effectively. But it took three days and the extent of the slowdown issue baffled even stack overflow. I've learned quite a few valuable lessons from this mess.

A few very minor bugs still exist but I will be fixing them in the future since they barely noticeable and cause no issues during runtime.