Hashmap and Observable list rework
Removed the methods used to create hashmaps and observable lists from the 'AddBooksController' and put them into the 'BookAttributes' class with the hashmaps and lists that they create.

This was done primarily to simplify the code, improve readability and greatly improving runtime performance by removing the need for global variables.

Additionally, many of the methods were turned into private methods.