Search Pane Controller Rework
Updated the query creation process to bring it in line with the new system.
Improved error handling
Removed unneeded variables, added new comments and cleaned up syntax.
Converted several methods to private.

More will be changed when the UI updates begin but this brings the current code up to date.