Fixed the stage size
Disabled the ability to resize the program stage to prevent breaking the UI. Things are in a fixed size and though dynamic elements may be considered in the future, resizing the current application merely breaks it.

The current size was specifically chosen to accommodate a 1920 x 1080 monitor in both portrait and landscape orientation.