Added a subtitle variable for books
Added a variable for book subtitles. This will allow books to be added with a subtitle instead of forcing it into the title field. This is much more in line with how books are treated in larger organizations and allows for a better look when looking at a book's information.