Several Changes
Removed all remnants of the "notification label" system that was used to indicated required fields that were not filled correctly. Those fields now change color to indicate the issue.

The fields no longer change color through Color Adjust and instead use -fx-background-color for the value. Still determining if this is better but initial results are promising.

Changed how the text fields are checked so that now each set of fields that are specific to certain book attributes can have their color adjusted and reset independently of each other. This will help to improve the process of adding multiple different attributes in an out of order approach should that prove necessary as it so often does during typical use cases.

Added subtitle string for new books.

Reworking the notification system to be cleaner and more easily understood.