Converted project to Modular
Converted the entire project over to a modular structure to bring it more in line with Maven standards. The project should no longer require adding any other libraries into the project structure or vm arguments. Along with cleaning up the layout slightly, this should make the process of actually getting the project up and running on different systems much easier along with deployments. This will be vital in the future.

There were also a long slew of bug fixes to make the code run with the new setup and there are still a small list that need to be fixed.