Changed how authors are created, stored and presented.
Authors used to have three separate fields for the individula parts of their names. This caused too much complexity and needless work so it has been simplified to only a single value. Now the name is one full String. The required changes have already been made to both the SQL table and stored procedures as well.