Now check for existing book attributes
The controller now checks if an attribute already exists in the hashmap specific to the new book attribute before attempting to add it. For example. When adding a new author, the program will first check if that author already exists before going forward with adding them.

The checks are not case sensitive, that may be a future change.